Create Balanced Parenthesis by using Stacks
This C++ program, using a stack data structure, computes whether the given parentheses expression is valid or not by checking whether each parenthesis is closed and nested in the input expression.
Here is the source code of the C++ program to display if it is a balanced expression or an invalid string. This C++ program is successfully compiled and run on DevCpp, a C++ compiler.