Reverse an array using pointers
In this program, we make use of * operator. 
In the reverse function, we take two pointers one pointing at the beginning of the array, other pointing at end of the array. The contents of the memory location pointed by these two pointers are swapped and then the value of the first pointer is increased and that of the second pointer is decreased.