Create Program to print the position of the smallest of  n numbers using array.
Using this program we can obtain the smallest number and where it is located in the series. The random numbers are taken by the user as an input and then the program finds the smallest number and its position from the input as an output. This program is made using C language.